Refactor marketplace to use int32 and int64 for quantities and prices
- Updated `mockBank` to use int32 for item quantities and int64 for flux amounts.
- Changed order and trade structures to use int32 for quantities and int64 for prices.
- Modified protobuf definitions to reflect changes in quantity and price types.
- Adjusted frontend code to handle new data types for flux balance and market operations.
- Ensured all tests are updated to accommodate the new integer types for consistency.

+++ internal/game/config.go
@@ -70,7 +70,7 @@ type GameConfig struct {
 	// Marketplace
 	MarketTaxPct      float64 `json:"marketTaxPct"`      // transaction tax (default 0.02 = 2%)
 	MarketOrderExpiry float64 `json:"marketOrderExpiry"` // hours until expiry (default 168 = 7 days)
-	MarketMinPrice    float64 `json:"marketMinPrice"`    // min price per unit (default 1)
+	MarketMinPrice    int64   `json:"marketMinPrice"`    // min price per unit (default 1)
 	MarketMaxOrders   int     `json:"marketMaxOrders"`   // max active orders per player (default 50)
 }
